tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/mathematical/RoundConverterTest.java
package com.xy.xsql.tsql.converter.functions.mathematical;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/5/31.
 */
public class RoundConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = RoundConverter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<ROUND> ::=\n" +
                        "ROUND ( numeric_expression , length [ , function ] )");
    }

}